Semaphore Solution to Producer-Consumer problem
Program starts with 100 emptys and 0 fulls.  The producer takes 1 from
the emptys and adds 1 to the fulls only if there is an empty, otherwise,
it waits.  The consumer takes from the fulls and adds to the emptys; if
there are no fulls, the consumer waits.  The producer and consumer are
each run ten times in the program.